requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
openai
jinja2
aiofiles